requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

contextily==1.6.2
fastapi==0.116.1
fastparquet==2024.11.0
geopandas==1.1.1
httpx==0.28.1
ipython==9.5.0
lxml==6.0.1
matplotlib==3.10.6
mysql-connector-python==9.4.0
numpy==2.3.3
pandas==2.3.2
plotly==6.3.0
prometheus-fastapi-instrumentator==7.1.0
pyarrow==21.0.0
pydantic==2.11.9
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
shapely==2.1.1
sqlalchemy==2.0.43
statsmodels==0.14.5
utm==0.8.1
uvicorn==0.30.1
wradlib==2.4.0
xarray==2025.9.0
xgboost==3.0.5
zarr==3.1.2
mlflow==3.4.0
python-dotenv==1.1.1
